requirements.txt
# Django C2 Platform - Requirements
# Production-ready Django setup for 10,000+ concurrent users

# Core Django
Django==5.0.1
djangorestframework==3.14.0

# Async & WebSocket (Django Channels)
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Database
psycopg2-binary==2.9.9  # PostgreSQL
mysqlclient==2.2.1      # MySQL (alternative)

# Celery (Async Task Queue)
celery==5.3.4
redis==5.0.1

# Cache & Session
django-redis==5.4.0

# Security & Auth
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
cryptography==41.0.7

# Performance
gunicorn==21.2.0
uvicorn[standard]==0.25.0
whitenoise==6.6.0

# Monitoring & Logging
django-debug-toolbar==4.2.0
sentry-sdk==1.39.2

# Utilities
python-decouple==3.8
django-extensions==3.2.3
django-filter==23.5

# Existing requirements
requests==2.31.0
psutil==5.9.6
Pillow==10.1.0
pyscreenshot==3.1

# Development
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.23.2
black==23.12.1
